Use MUnit in WriteControllerImplSuite
